Moved ToolControlFlowSignal to core
This commit moves ToolControlFlowSignal to core, to resolve a cyclic
dependency.

See: gh-1024
