+++ embabel-agent-api/src/main/kotlin/com/embabel/agent/spi/common/RetryProperties.kt
@@ -15,7 +15,7 @@
  */
 package com.embabel.agent.spi.common
 
-import com.embabel.agent.api.tool.ToolControlFlowSignal
+import com.embabel.agent.core.ToolControlFlowSignal
 import com.embabel.agent.spi.support.LlmDataBindingProperties.Companion.isRateLimitError
 import com.embabel.agent.spi.support.springai.SpringAiRetryPolicy
 import com.embabel.common.util.loggerFor
